Setup react-navigation with routing on web
Working navigation setup for web + native. NextJS page routing works differently
than react-navigation, but we can handle that with expo-next-react-navigation
to control how we navigate on each platform. We also might just ditch NextJS
entirely for most of the app if the expo generated PWA is sufficient, as they have
a pretty good setup for handling URLs and Deeplinks in a unified way.